clean_with_gemma_api/gemini_api_orchestrator/requirements.txt
google-generativeai>=0.7.0
boto3>=1.35.0
pandas>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0